requirements.txt
streamlit==1.31.0
openai==0.28
pandas==2.2.0
python-dotenv==1.0.0 